+++ src/Quadrupen/Regularizer.h
@@ -20,9 +20,9 @@ public:
   Regularizer(const RegressionData<matrix>&, const List&);
 
   RegressionData<matrix> data_ ; // data structure
+  double gamma_                ; // overall amount of minor penalty (not leading the path)
   vector<double> lambdas_      ; // vector of parameters tuning the main penalty
   vec lambda_factor_           ; // main penalty factors
-  double gamma_                ; // overall amount of minor penalty (not leading the path)
   vector<double> intercept_    ; // vector of intercept term
   matrix coef_                 ; // matrix of coefficients
   vec beta_                    ; // vector of current parameters
