+++ src/wrapper_GroupLasso.cpp
@@ -3,13 +3,13 @@
  *         MIA Paris-Saclay
  */
 
-#include "Quadrupen/GroupLasso.h"
+#include "Quadrupen/GroupElasticNet.h"
 
 using namespace Rcpp;
 using namespace arma;
 
 // [[Rcpp::export]]
-List group_lasso_l1l2_dense_cpp(
+List group_enet_l1l2_dense_cpp(
     const Environment &dataModel, // data structure
     const bool        &intercept, // Boolean for intercept
     const arma::uvec  &group    , // Vector of groups
@@ -19,29 +19,29 @@ List group_lasso_l1l2_dense_cpp(
   
   RegressionData<mat> data(dataModel, intercept, as<bool>(control["normalize"])) ;
 
-  GroupElasticNet<mat> grplasso(data, group, regParam, control);
+  GroupElasticNet<mat> grpenet(data, group, regParam, control);
   
-  List results = grplasso.solution_path(control);
+  List results = grpenet.solution_path(control);
 
   return List::create(
     Named("tuning_param") = List::create(
-      Named("l1") = grplasso.path_tuning(),
-      Named("l2") = grplasso.gamma_
+      Named("l1") = grpenet.path_tuning(),
+      Named("l2") = grpenet.gamma_
     ),
-    Named("coef")               = grplasso.coefficients(),
-    Named("coef_debiased")      = grplasso.debiased_coefficients(),
-    Named("active")             = grplasso.active_var(),
-    Named("intercept")          = grplasso.intercept_,
-    Named("intercept_debiased") = grplasso.intercept_debiased_,
-    Named("normx")              = grplasso.data_.norm_X_,
-    Named("df")                 = grplasso.df_,
+    Named("coef")               = grpenet.coefficients(),
+    Named("coef_debiased")      = grpenet.debiased_coefficients(),
+    Named("active")             = grpenet.active_var(),
+    Named("intercept")          = grpenet.intercept_,
+    Named("intercept_debiased") = grpenet.intercept_debiased_,
+    Named("normx")              = grpenet.data_.norm_X_,
+    Named("df")                 = grpenet.df_,
     Named("monitoring")         = results
   );
   
 }
 
 // [[Rcpp::export]]
-List group_lasso_l1l2_sparse_cpp(
+List group_enet_l1l2_sparse_cpp(
     const Environment &dataModel, // data structure
     const bool        &intercept, // Boolean for intercept
     const arma::uvec  &group    , // Vector of groups
@@ -51,22 +51,22 @@ List group_lasso_l1l2_sparse_cpp(
   
   RegressionData<sp_mat> data(dataModel, intercept, as<bool>(control["normalize"])) ;
   
-  GroupElasticNet<sp_mat> grplasso(data, group, regParam, control);
+  GroupElasticNet<sp_mat> grpenet(data, group, regParam, control);
   
-  List results = grplasso.solution_path(control);
+  List results = grpenet.solution_path(control);
   
   return List::create(
     Named("tuning_param") = List::create(
-      Named("l1") = grplasso.path_tuning(),
-      Named("l2") = grplasso.gamma_
+      Named("l1") = grpenet.path_tuning(),
+      Named("l2") = grpenet.gamma_
     ),
-    Named("coef")               = grplasso.coefficients(),
-    Named("coef_debiased")      = grplasso.debiased_coefficients(),
-    Named("active")             = grplasso.active_var(),
-    Named("intercept")          = grplasso.intercept_,
-    Named("intercept_debiased") = grplasso.intercept_debiased_,
-    Named("normx")              = grplasso.data_.norm_X_,
-    Named("df")                 = grplasso.df_,
+    Named("coef")               = grpenet.coefficients(),
+    Named("coef_debiased")      = grpenet.debiased_coefficients(),
+    Named("active")             = grpenet.active_var(),
+    Named("intercept")          = grpenet.intercept_,
+    Named("intercept_debiased") = grpenet.intercept_debiased_,
+    Named("normx")              = grpenet.data_.norm_X_,
+    Named("df")                 = grpenet.df_,
     Named("monitoring")         = results
   );
